Refactor: split into disasm.rs
